Refactor: move default injectors and API solvers under inject
