+++ brevitas/nn/quant_avg_pool.py
@@ -48,11 +48,10 @@ from dependencies import Injector
 
 from brevitas.function.ops_ste import ceil_ste
 from brevitas.function.ops import max_uint
-
 from brevitas.proxy.runtime_quant import AccQuantProxyProtocol
-from brevitas.proxy.config import update_trunc_quant_injector
 from brevitas.quant_tensor import QuantTensor
-from .quant_layer import DefaultTruncQuantInjector as DefaultTruncQI
+from brevitas.inject.solver import update_trunc_quant_injector
+from brevitas.inject.defaults import DefaultTruncQuantInjector as DefaultTruncQI
 from .mixin.base import QuantLayerMixin
 from .mixin.acc import QuantTruncMixin
