+++ brevitas/nn/quant_bn.py
@@ -46,8 +46,8 @@ from dependencies import Injector
 import brevitas.config as config
 from brevitas.proxy.parameter_quant import WeightQuantProxyProtocol, BiasQuantProxyProtocol
 from brevitas.proxy.runtime_quant import ActQuantProxyProtocol
-from .quant_layer import DefaultWeightQuantInjector as DefaultWeightQI
-from .quant_layer import DefaultBiasQuantInjector as DefaultBiasQI
+from brevitas.inject.defaults import DefaultWeightQuantInjector as DefaultWeightQI
+from brevitas.inject.defaults import DefaultBiasQuantInjector as DefaultBiasQI
 from .quant_scale_bias import QuantScaleBias
 from .utils import mul_add_from_bn
